% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_files_grm.R
\name{require_files_grm}
\alias{require_files_grm}
\title{Require that GCTA binary GRM files are present}
\usage{
require_files_grm(file)
}
\arguments{
\item{file}{The shared file path (excluding extensions: grm.bin, grm.N.bin, or grm.id).}
}
\value{
Nothing
}
\description{
This function checks that each of the GCTA binary GRM files (grm.bin, grm.N.bin, and grm.id extensions) are present, given the shared base file path, stopping with an informative message if any of the files is missing.
This function aids troubleshooting, as various downstream external software report missing files differently and sometimes using confusing or obscure messages.
}
\examples{
# check that the samples we want exist
# start with bed file
file <- system.file("extdata", 'sample.grm.bin', package = "genio", mustWork = TRUE)
# remove extension
file <- sub('\\\\.grm\\\\.bin$', '', file)
# since all sample.grm.{bin,N.bin,id} files exist, this will not stop with error messages:
require_files_grm(file)

}
