% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_grm.R
\name{delete_files_grm}
\alias{delete_files_grm}
\title{Delete all GCTA binary GRM files}
\usage{
delete_files_grm(file)
}
\arguments{
\item{file}{The shared file path (excluding extensions: grm.bin, grm.N.bin, or grm.id).}
}
\value{
Nothing
}
\description{
This function deletes each of the GCTA binary BRM files (grm.bin, grm.N.bin, and grm.id extensions) given the shared base file path, warning with an informative message if any of the files did not exist.
}
\examples{
# create dummy GRM files
file <- tempfile('delete-me-test') # no extension
# add each extension and create empty files
file.create( paste0(file, '.grm.bin') )
file.create( paste0(file, '.grm.N.bin') )
file.create( paste0(file, '.grm.id') )

# delete the GRM files we just created
delete_files_grm(file)

}
