% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_phen.R
\name{delete_files_phen}
\alias{delete_files_phen}
\title{Delete PHEN files}
\usage{
delete_files_phen(file)
}
\arguments{
\item{file}{The base file path (excluding PHEN extensions).}
}
\value{
Nothing
}
\description{
This function deletes a PHEN files given the base file path (without extension), warning with an informative message if the file did not exist.
}
\examples{
# create dummy PHEN files
file <- tempfile('delete-me-test') # no extension
# add extension and create an empty file
file.create( paste0(file, '.phen') )

# delete the PHEN file we just created
delete_files_phen(file)

}
