% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eigenvec.R
\name{read_eigenvec}
\alias{read_eigenvec}
\title{Read plink eigenvec file}
\usage{
read_eigenvec(file, ext = "eigenvec", verbose = TRUE)
}
\arguments{
\item{file}{The input file path, potentially excluding extension.}

\item{ext}{File extension (default "eigenvec") can be changed if desired.}

\item{verbose}{If TRUE (default) function reports the path of the file being written (after autocompleting the extension).}
}
\value{
A list with two elements:
\itemize{
\item \code{eigenvec}: A numeric R matrix containing the parsed eigenvectors
\item \code{fam}: A tibble with two columns, \code{fam} and \code{id}, which are the first two columns of the parsed file.
}
}
\description{
This function reads a plink eigenvec file, parsing columns strictly.
First two must be 'fam' and 'id', which are strings, and all remaining columns (eigenvectors) must be numeric.
}
\examples{
# read an existing *.eigenvec file
file <- system.file("extdata", 'sample.eigenvec', package = "genio", mustWork = TRUE)
data <- read_eigenvec(file)
# numeric eigenvector matrix
data$eigenvec
# fam/id tibble
data$fam

# can specify without extension
file <- sub('\\\\.eigenvec$', '', file) # remove extension from this path on purpose
file # verify .eigenvec is missing
data <- read_eigenvec(file) # load it anyway!
data$eigenvec

}
\seealso{
\code{\link{write_eigenvec}} for writing an eigenvec file.

Plink 1 eigenvec format reference:
\url{https://www.cog-genomics.org/plink/1.9/formats#eigenvec}

Plink 2 eigenvec format reference:
\url{https://www.cog-genomics.org/plink/2.0/formats#eigenvec}

GCTA eigenvec format reference:
\url{https://cnsgenomics.com/software/gcta/#PCA}
}
