% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bed.R
\name{write_bed}
\alias{write_bed}
\title{Write a genotype matrix into plink BED format}
\usage{
write_bed(file, X)
}
\arguments{
\item{file}{Output file path.  .bed extension may be omitted (will be added automatically if it is missing).}

\item{X}{The \eqn{m \times n}{m-by-n} genotype matrix.}
}
\value{
Nothing
}
\description{
This function accepts a standard R matrix containing genotypes (values in \code{c(0,1,2,NA)}) and writes it into a plink-formatted BED (binary) file.
Each genotype per locus (m loci) and individual (n total) counts the number of alternative alleles or \code{NA} for missing data.
}
